+++ CHANGELOG.md
@@ -7,7 +7,8 @@
   For compressed files, this limit applies to the _decompressed_ size.
   The default value is 15GiB. ([#1993](https://github.com/getsentry/symbolicator/pull/1993))
 
-- Disabled per-request auto-decompression on the `trusted` download client.
+- Introduced a `sentry` download client used for Sentry file downloads, which is
+  analogous to the `trusted` client, but disables automatic decompression.
   This allows Symbolicator to support compressed files served by Sentry.
   ([#1999](https://github.com/getsentry/symbolicator/pull/1999))