+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Support compressed range requests. This allows backends (like S3 and objectstore) to serve partial
   ranges from a compressed file. ([#2004](https://github.com/getsentry/symbolicator/pull/1999))
 - Raised the default of `object_file_max_decompressed_source_size` from 100MiB to 190MiB. ([#2009](https://github.com/getsentry/symbolicator/pull/2009))
+- Use IO-Uring for file operations, if it is available. ([#2014](https://github.com/getsentry/symbolicator/pull/2014))
 
 ## 26.7.2